+++ exp_by_day/10.14/exp07j:full_KS_without_downsample_resume_5_f.sh
@@ -5,5 +5,8 @@
 
 # bash ./scripts/train.sh -b 2 -c "2,3,4,5,6,7" -g 42 -w Exp07j -R /data1/lcy/projects/ReCamMaster/wandb/10-08-111048_Exp07g/checkpoints/step2722.ckpt
 
+# ## 因为训练中断，接着上一次训练的继续
+# bash ./scripts/train.sh -b 2 -c "2,3,4,5,6,7" -g 42 -w Exp07j -R /data1/lcy/projects/ReCamMaster/wandb/10-14-134124_Exp07j/checkpoints/step300.ckpt
+
 ## 因为训练中断，接着上一次训练的继续
-bash ./scripts/train.sh -b 2 -c "2,3,4,5,6,7" -g 42 -w Exp07j -R /data1/lcy/projects/ReCamMaster/wandb/10-14-134124_Exp07j/checkpoints/step300.ckpt
+bash ./scripts/train.sh -b 2 -c "4,5,6,7" -g 42 -w Exp07j -R /data1/lcy/projects/ReCamMaster/wandb/10-15-095226_Exp07j/checkpoints/step100.ckpt